// src/app/model/rk_hc_form_cab.ts
export class rk_hc_form_cab {
    id: number;
    hacienda_id: number;
    lote: string;
    modulo: string;
    empresa_id:number;
    tipo_muestra_id: number;
    observacion: string;
    estado: string;
    usuario_cre_id: number;
    fecha_cre: string;
    usuario_mod_id: number;
    fecha_mod: string;
}